pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "symplastic-growth"
version = "0.1.0"
description = "Library for modeling symplastic leaf growth: osmosis, turgor, cell wall mechanics, cell division, and optional surrogate modeling."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Ulyana Zubairova", email = "u.zubairova@g.nsu.ru" },
]
keywords = [
    "plant growth",
    "symplastic",
    "turgor",
    "osmosis",
    "cell division",
    "modeling",
    "surrogate",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "scikit-learn>=1.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.5"]
jupyter = ["jupyter>=1.0", "matplotlib>=3.5"]
# YAML configs (load_params / save_params with .yaml)
config = ["pyyaml>=5.4"]
dev = [
    "pytest>=7.0",
    "jupyter>=1.0",
    "matplotlib>=3.5",
    "pyyaml>=5.4",
]

[project.urls]
Homepage = "https://github.com/zubairovaus/symplastic-growth"
Repository = "https://github.com/zubairovaus/symplastic-growth"
Documentation = "https://github.com/zubairovaus/symplastic-growth#readme"
"Bug Tracker" = "https://github.com/zubairovaus/symplastic-growth/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["symplastic_growth*", "surrogate*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
filterwarnings = [
    "ignore::UserWarning",
]
